Switch data type from double to decimal

+++ Reference Implementation/OANDARestLibrary/TradeLibrary/DataTypes/Candle.cs
@@ -4,33 +4,33 @@
     {
         #region Public Properties
 
-        public double closeAsk { get; set; }
+        public  decimal closeAsk { get; set; }
 
-        public double closeBid { get; set; }
+        public  decimal closeBid { get; set; }
 
-        public double closeMid { get; set; }
+        public  decimal closeMid { get; set; }
 
         public bool complete { get; set; }
 
-        public double highAsk { get; set; }
+        public  decimal highAsk { get; set; }
 
         // Bid/Ask candles
 
-        public double highBid { get; set; }
+        public  decimal highBid { get; set; }
 
-        public double highMid { get; set; }
+        public  decimal highMid { get; set; }
 
-        public double lowAsk { get; set; }
+        public  decimal lowAsk { get; set; }
 
-        public double lowBid { get; set; }
+        public  decimal lowBid { get; set; }
 
-        public double lowMid { get; set; }
+        public  decimal lowMid { get; set; }
 
-        public double openAsk { get; set; }
+        public  decimal openAsk { get; set; }
 
-        public double openBid { get; set; }
+        public  decimal openBid { get; set; }
 
-        public double openMid { get; set; }
+        public  decimal openMid { get; set; }
 
         public string time { get; set; }
